+++ metrics/utils.py
@@ -185,7 +185,11 @@ def get_verbose_name(class_name):
 def request_is_ajax(request):
     return (
         request.META.get("HTTP_X_REQUESTED_WITH") == "XMLHttpRequest"
-        or request.META.get("HTTP_HX_REQUEST") == "true"  # htmx
+        # Htmx.
+        or (
+            request.META.get("HTTP_HX_REQUEST") == "true"
+            and request.META.get("HTTP_HX_BOOSTED") != "true"  # Do not ignore boosted htmx requests.
+        )
     )